improved namespace management
