+++ pliers/converters/multistep.py
@@ -1,6 +1,6 @@
 from pliers.converters import Converter
 from pliers.converters import get_converter
-from pliers.stimuli import Stim
+from pliers.stimuli.base import Stim
 from pliers.stimuli.audio import AudioStim
 from pliers.stimuli.video import VideoStim
 from pliers.stimuli.text import TextStim
