+++ pliers/extractors/text.py
@@ -3,7 +3,7 @@ Extractors that operate primarily or exclusively on Text stimuli.
 '''
 
 from pliers.stimuli.text import TextStim, ComplexTextStim
-from pliers.extractors import Extractor, ExtractorResult
+from pliers.extractors.base import Extractor, ExtractorResult
 from pliers.support.exceptions import PliersError
 from pliers.support.decorators import requires_nltk_corpus
 from pliers.datasets.text import fetch_dictionary, datasets
