+++ pliers/tests/test_transformers.py
@@ -1,8 +1,8 @@
 from pliers.transformers import get_transformer
-from pliers.extractors import Extractor
+from pliers.extractors.base import Extractor
 from pliers.extractors.audio import STFTAudioExtractor
 from pliers.tests.utils import get_test_data_path, DummyExtractor
-from pliers.stimuli import TransformationLog
+from pliers.stimuli.base import TransformationLog
 from pliers.stimuli.image import ImageStim
 from os.path import join
